[fs] btrfs: remove some no-op casts
Message-id: <20170307000403.5752-457-billodo@redhat.com>
Patchwork-id: 171641
O-Subject: [RHEL7.4 PATCH BZ 1399590 456/485] Btrfs: remove some no-op casts
Bugzilla: 1399590
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

Bugzilla: 1399590
Upstream Status: Committed upstream
Build Info: https://brewweb.devel.redhat.com/taskinfo?taskID=12695624
Tested: Locally with fstests

Backport of the following upstream commit:

    commit 9c894696f56f5d84fb5766af81bcda4a7cb9a842
    Author: Dan Carpenter <dan.carpenter@oracle.com>
    Date:   Wed Oct 12 11:33:21 2016 +0300

    Btrfs: remove some no-op casts

    We cast 0 to a u8 but then because of type promotion, it's immediately
    cast to int back to int before we do a bitwise negate.  The cast doesn't
    matter in this case, the code works as intended.  It causes a static
    checker warning though so let's remove it.

+++ fs/btrfs/extent_io.c
@@ -5577,7 +5577,7 @@ void le_bitmap_set(u8 *map, unsigned int start, int len)
 		*p |= mask_to_set;
 		len -= bits_to_set;
 		bits_to_set = BITS_PER_BYTE;
-		mask_to_set = ~(u8)0;
+		mask_to_set = ~0;
 		p++;
 	}
 	if (len) {
@@ -5597,7 +5597,7 @@ void le_bitmap_clear(u8 *map, unsigned int start, int len)
 		*p &= ~mask_to_clear;
 		len -= bits_to_clear;
 		bits_to_clear = BITS_PER_BYTE;
-		mask_to_clear = ~(u8)0;
+		mask_to_clear = ~0;
 		p++;
 	}
 	if (len) {
@@ -5687,7 +5687,7 @@ void extent_buffer_bitmap_set(struct extent_buffer *eb, unsigned long start,
 		kaddr[offset] |= mask_to_set;
 		len -= bits_to_set;
 		bits_to_set = BITS_PER_BYTE;
-		mask_to_set = ~(u8)0;
+		mask_to_set = ~0;
 		if (++offset >= PAGE_CACHE_SIZE && len > 0) {
 			offset = 0;
 			page = eb->pages[++i];
@@ -5729,7 +5729,7 @@ void extent_buffer_bitmap_clear(struct extent_buffer *eb, unsigned long start,
 		kaddr[offset] &= ~mask_to_clear;
 		len -= bits_to_clear;
 		bits_to_clear = BITS_PER_BYTE;
-		mask_to_clear = ~(u8)0;
+		mask_to_clear = ~0;
 		if (++offset >= PAGE_CACHE_SIZE && len > 0) {
 			offset = 0;
 			page = eb->pages[++i];
